PRD: Update PRD simulator to support 64-bit
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101236
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101261
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/usr/diag/prdf/common/framework/service/prdfServiceDataCollector.C
@@ -393,7 +393,7 @@ uint32_t ServiceDataCollector::Flatten(uint8_t * i_buffer, uint32_t & io_size) c
         buffer_append(current_ptr,l_huid);
         buffer_append(current_ptr,error_signature.getSigId());
         // callouts
-        buffer_append(current_ptr,xMruList.size());
+        buffer_append(current_ptr, (uint32_t)xMruList.size());
         for ( SDC_MRU_LIST::const_iterator i = xMruList.begin();
               i != xMruList.end(); ++i )
         {
@@ -403,7 +403,7 @@ uint32_t ServiceDataCollector::Flatten(uint8_t * i_buffer, uint32_t & io_size) c
             buffer_append( current_ptr, (uint32_t)i->gardState         );
         }
 
-        buffer_append(current_ptr, iv_SignatureList.size());
+        buffer_append(current_ptr,  (uint32_t)iv_SignatureList.size());
         for(PRDF_SIGNATURES::const_iterator i = iv_SignatureList.begin();
             i != iv_SignatureList.end(); ++i)
         {
